+++ CHANGELOG.md
@@ -3,6 +3,9 @@
  - 2020-xx-xx
 
 ### Changes since the last release
+ 
+ - NavBall target and crash markers have been fixed and updated,
+   target is now a green circle and predicted crash site is now a red square.
  - Due to performance issues with larger craft, the Cache is now enabled by default. For higher precision,
    it can still be turned on manually.